[#265] Integrate App Rating Alert
- update of SPM

[#265] Integrate App Rating Alert

- async adaptations of the latest sdk
- review request TCA dependency added
- set of rules for the triggering the app rating dialog
- deeplinks commented out for now
- WIP for the tests

[#265] Integrate App Rating Alert

- unit tests fixed
- deeplinks fixed

[#265] Integrate App Rating Alert

- unit tests for the review request client
- app review request logic disconnected from the production for now

[#265] Integrate App Rating Alert (#704)

- added the TODO for the triggering

+++ secant/Features/Root/RootDestination.swift
@@ -141,7 +141,7 @@ private extension RootReducer {
         case .home:
             return .destination(.deeplinkHome)
         case let .send(amount, address, memo):
-            return .destination(.deeplinkSend(Zatoshi(amount), address, memo))
+            return .destination(.deeplinkSend(Zatoshi(Int64(amount)), address, memo))
         }
     }
 }
